move HTTP_STATUS... to consts
