+++ src/consts.js
@@ -1,0 +1,3 @@
+export const HTTP_STATUS_NOT_FOUND = 404
+export const HTTP_STATUS_UNAUTHORIZED = 401
+export const HTTP_STATUS_FORBIDDEN = 403
